feat: attribute segment reactions to the exact quoted part in prompts
