+++ CHANGELOG.md
@@ -6,7 +6,7 @@ This file is the release-notes source of truth for Marinara Engine. Reuse these 
 
 ### Added
 
-- Conversation mode: emoji reactions can now target an individual character's part of a merged multi-character reply. Each speaker segment gets its own add-reaction button (next to the name) and its own reaction row, and characters are told whose part a reaction was aimed at; reactions without a segment target keep applying to the whole message, so existing chats and 1:1 conversations are unaffected (#3210).
+- Conversation mode: emoji reactions can now target an individual character's part of a merged multi-character reply. Each speaker segment gets its own add-reaction button (next to the name) and its own reaction row, and characters are told exactly which part a reaction was aimed at (the targeted part is quoted back to them); reactions without a segment target keep applying to the whole message, so existing chats and 1:1 conversations are unaffected (#3210).
 
 - Added the Agent Suite to the Chat Settings drawer's Agents section: a window listing the agents active in the current chat where you can view and edit everything they have stored — agent memory, tracker state, and custom-agent outputs — manually or with AI-assisted rewrites (select text, give an instruction, optionally attach grounding context such as character cards or active-lorebook entries, and pick a connection) (#3160).
 - Added first-class scene video generation for Game Mode, Roleplay, and Visual Novel galleries, including Video Generation connections for Gemini Omni and xAI Imagine, editable `game.video` prompts, manual Gallery video actions, per-image Animate buttons, Gallery video previews with prompt copy, live View Latest media, and draggable/resizable pinned video overlays.
